+++ os/StarryOS/kernel/src/config/riscv64.rs
@@ -4,10 +4,10 @@ pub const KERNEL_STACK_SIZE: usize = 0x4_0000;
 /// The base address of the user space.
 pub const USER_SPACE_BASE: usize = 0x1000;
 /// The size of the user space.
-pub const USER_SPACE_SIZE: usize = 0x3f_ffff_f000;
+pub const USER_SPACE_MAX_SIZE: usize = 0x3f_ffff_f000;
 
 /// The highest address of the user stack.
-pub const USER_STACK_TOP: usize = 0x4_0000_0000;
+pub const USER_STACK_TOP_MAX: usize = 0x4_0000_0000;
 /// The size of the user stack.
 pub const USER_STACK_SIZE: usize = 0x80_0000;