+++ os/StarryOS/kernel/src/mm/mod.rs
@@ -3,8 +3,9 @@
 mod access;
 mod aspace;
 mod io;
+mod layout;
 mod loader;
 mod stats;
 mod vm_stat;
 
-pub use self::{access::*, aspace::*, io::*, loader::*, stats::*, vm_stat::*};
+pub use self::{access::*, aspace::*, io::*, layout::*, loader::*, stats::*, vm_stat::*};